move multimap to its own file, with tests
